+++ bench-vortex/src/utils/logging.rs
@@ -1,20 +1,41 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
+use std::fs::File;
 use std::io::IsTerminal;
 
 use tracing::level_filters::LevelFilter;
+use tracing_perfetto::PerfettoLayer;
 use tracing_subscriber::EnvFilter;
+use tracing_subscriber::prelude::*;
 
-pub fn setup_logger(filter: EnvFilter) {
-    tracing_subscriber::fmt()
+/// Initialize logging/tracing for a benchmark
+pub fn setup_logging_and_tracing(verbose: bool, tracing: bool) -> anyhow::Result<()> {
+    let filter = default_env_filter(verbose);
+
+    let fmt_layer = tracing_subscriber::fmt::layer()
         .with_writer(std::io::stderr)
-        .with_file(true)
         .with_level(true)
+        .with_file(true)
         .with_line_number(true)
-        .with_env_filter(filter)
-        .with_ansi(std::io::stderr().is_terminal())
+        .with_ansi(std::io::stderr().is_terminal());
+
+    tracing_subscriber::registry()
+        .with(filter)
+        .with(
+            tracing
+                .then(|| {
+                    Ok::<_, anyhow::Error>(
+                        PerfettoLayer::new(File::create("trace.json")?)
+                            .with_debug_annotations(true),
+                    )
+                })
+                .transpose()?,
+        )
+        .with(fmt_layer)
         .init();
+
+    Ok(())
 }
 
 pub fn default_env_filter(is_verbose: bool) -> EnvFilter {
