+++ vortex-datafusion/Cargo.toml
@@ -25,11 +25,8 @@ moka = { workspace = true, features = ["future"] }
 object_store = { workspace = true }
 tokio = { workspace = true, features = ["rt-multi-thread", "fs"] }
 tokio-stream = { workspace = true }
-tracing = { workspace = true, optional = true }
+tracing = { workspace = true }
 vortex = { workspace = true, features = ["object_store", "tokio"] }
-
-[features]
-tracing = ["dep:tracing", "vortex/tracing"]
 
 [dev-dependencies]
 anyhow = { workspace = true }
