+++ vortex-file/Cargo.toml
@@ -29,7 +29,7 @@ moka = { workspace = true, features = ["future"] }
 object_store = { workspace = true, optional = true }
 rustc-hash = { workspace = true }
 tokio = { workspace = true, features = ["rt"], optional = true }
-tracing = { workspace = true, optional = true }
+tracing = { workspace = true }
 # Needed to pickup the "js" feature for wasm targets from the workspace configuration
 uuid = { workspace = true }
 vortex-alp = { workspace = true }
@@ -81,5 +81,4 @@ tokio = [
     "vortex-layout/tokio",
     "vortex-scan/tokio",
 ]
-tracing = ["dep:tracing", "vortex-io/tracing", "vortex-layout/tracing"]
 zstd = ["dep:vortex-zstd", "vortex-layout/zstd"]
