+++ vortex-io/src/tokio.rs
@@ -50,10 +50,7 @@ impl Deref for TokioFile {
 }
 
 impl VortexReadAt for TokioFile {
-    #[cfg_attr(
-        feature = "tracing",
-        tracing::instrument(skip_all, fields(range, alignment))
-    )]
+    #[tracing::instrument(skip_all, fields(range, alignment))]
     async fn read_byte_range(
         &self,
         range: Range<u64>,
@@ -75,7 +72,7 @@ impl VortexReadAt for TokioFile {
         PerformanceHint::local()
     }
 
-    #[cfg_attr(feature = "tracing", tracing::instrument(skip_all))]
+    #[tracing::instrument(skip_all)]
     async fn size(&self) -> io::Result<u64> {
         self.metadata().map(|metadata| metadata.len())
     }
