+++ vortex-layout/Cargo.toml
@@ -32,7 +32,7 @@ pin-project-lite = { workspace = true }
 prost = { workspace = true }
 roaring = { workspace = true, optional = true }
 tokio = { workspace = true, features = ["rt"], optional = true }
-tracing = { workspace = true, optional = true }
+tracing = { workspace = true }
 vortex-array = { workspace = true }
 vortex-btrblocks = { workspace = true }
 vortex-buffer = { workspace = true }
@@ -57,7 +57,7 @@ vortex-layout = { path = ".", features = ["tokio", "test-harness"] }
 [features]
 roaring = ["dep:roaring"]
 test-harness = []
-tokio = ["dep:tokio", "vortex-error/tokio", "tracing"]
+tokio = ["dep:tokio", "vortex-error/tokio"]
 zstd = ["dep:vortex-zstd"]
 
 [lints]
